add isset

+++ includes/pages/adm/ShowActivePage.php
@@ -21,8 +21,11 @@ function ShowActivePage()
 {
 	global $LNG, $USER;
 	$id = HTTP::_GP('id', 0);
-	if($_GET['action'] == 'delete' && !empty($id))
-		$GLOBALS['DATABASE']->query("DELETE FROM ".USERS_VALID." WHERE `validationID` = '".$id."' AND `universe` = '".Universe::getEmulated()."';");
+	if(isset($_GET['action'])) {
+		if($_GET['action'] == 'delete' && !empty($id)) {
+			$GLOBALS['DATABASE']->query("DELETE FROM ".USERS_VALID." WHERE `validationID` = '".$id."' AND `universe` = '".Universe::getEmulated()."';");
+		}
+	}
 
 	$query = $GLOBALS['DATABASE']->query("SELECT * FROM ".USERS_VALID." WHERE `universe` = '".Universe::getEmulated()."' ORDER BY validationID ASC");
